Patch around TCP_NODELAY bug.
There is a bug were not having TCP_NODELAY set causes connection issues
with Murano.  While ultimatily the bug is up there, we need to work
around it down here.  As for Ruby, setting TCP_NODELAY was added in 2.1.
But since the default version installed on macos is 2.0.0 we hit it.

So, if the current version of Ruby is 2.0.0, then use this bit of code I
copied from Ruby 2.1.